Added MQTT-C library (#9)
* Added MQTT-C submodule

* Added mqtt to build system

* chore: Add mqtt_module to retransmitter project

* feat: Update retransmitter configuration loading

The code changes in this commit update the configuration loading process for the retransmitter. The `load_configuration` function now accepts a `retransmitter_configuration_t` struct instead of `nrf24_configuration_t`. Additionally, the function has been modified to load the MQTT broker and port from the configuration file and environment variables.

* Implemented mqtt connection to broker

* feat: Add mqtt_module to retransmitter project

* Added module init and some logs

* feat: Add stdlib.h header to mqtt_module.h and added message receive and publish

* feat: Add message received callback to mqtt_module_init

The code changes in this commit modify the `mqtt_module_init` function in `mqtt_module.c` to accept a callback function for handling received MQTT messages. This allows the module to notify the application when a message is received on a subscribed topic.